Fix issues cause by migration V25 and V26
Fixes #6017

Please note that migrating sandbox-classic from version 1.0.0 to 1.1.1 will cause
undefined behavior if the result is used to back a running ledger setup.

No migration should end at 1.1.1 and they should move past that until the version
that applies this fix to make sure the index database is in a consistent state.

Full write-up on the issue here: https://github.com/digital-asset/daml/issues/6017#issuecomment-634040972

changelog_begin
[Sandbox Classic] Fix issue in migrating to version 1.1.1. If you did migrate to
version 1.1.1, do not use the resulting setup but make sure to migrate until this
version, which fixes the issue
See https://github.com/digital-asset/daml/issues/6017
changelog_end
