+++ ledger/sandbox/src/main/resources/db/migration/postgres/V32_0__Drop_archived_contracts.sha256
@@ -1,0 +1,1 @@
+b72887c806ad58d25e4678134f44e2410f3d98838106e2e8a91f98fbf46dce27
